requirements.txt
arcgis
geopandas
python-dotenv
sqlalchemy
psycopg2-binary

